[Ref] libopenmpt: Silence warnigns in examples on Win64.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5147 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ examples/libopenmpt_example_c_mem.c
@@ -55,7 +55,7 @@ int main( int argc, char * argv[] ) {
 		if ( count == 0 ) {
 			break;
 		}
-		Pa_WriteStream( stream, buffers, count );
+		Pa_WriteStream( stream, buffers, (unsigned long)count );
 	}
 	Pa_StopStream( stream );
 	Pa_CloseStream( stream );
